refactor: connect to db
