+++ product-service/package.json
@@ -8,6 +8,7 @@
     "fix": "eslint . --ext .ts --fix",
     "test": "jest",
     "service:deploy": "sls deploy",
+    "service:offline": "sls offline",
     "typeorm": " ./node_modules/typeorm/cli.js --config src/database/connection-options.ts",
     "typeorm:cli": "ts-node ./node_modules/typeorm/cli -config src/database/connection-options.ts",
     "migration:generate": "./node_modules/.bin/ts-node ./node_modules/.bin/typeorm migration:generate -d src/database/connection-options.ts",
@@ -39,6 +40,7 @@
     "json-schema-to-ts": "^1.5.0",
     "prettier": "^2.7.1",
     "serverless": "^3.0.0",
+    "serverless-dotenv-plugin": "^4.0.2",
     "serverless-offline": "^11.0.1",
     "serverless-webpack": "^5.3.5",
     "ts-jest": "^29.0.2",
