+++ product-service/serverless.ts
@@ -17,7 +17,11 @@ const serverlessConfiguration: AWS = {
       includeModules: { forceInclude: ['pg'] },
     },
   },
-  plugins: ['serverless-webpack', 'serverless-offline'],
+  plugins: [
+    'serverless-webpack',
+    'serverless-offline',
+    'serverless-dotenv-plugin',
+  ],
   provider: {
     name: 'aws',
     runtime: 'nodejs14.x',
