+++ product-service/src/functions/get-product-by-id/handler.ts
@@ -1,12 +1,11 @@
 import { middyfy } from '@libs/lambda';
 import { sendCustomResponse, sendError } from '../../utils/responses';
-import { filmService } from '../../film-service';
+import { filmService } from '../../services/film-service';
 
 export const getFilmById = async (event) => {
   try {
     const filmId = await event.pathParameters.id;
     const film = await filmService.getFilmById(filmId);
-    console.log(`Received film: ${JSON.stringify(film)}`);
     if (!film) {
       return sendCustomResponse({ message: 'Film not found' }, 404);
     }
